Move project metadata/config to pyproject.toml
We also remove both the setup.cfg and setup.py files as they are not
required by newer versions of pip and setuptools (but for this we also
need to bump the setuptools version).

If you are getting this error locally after the upgrade:

  ERROR: Project file:///home/luca/devel/frequenz-sdk-python has
  a 'pyproject.toml' and its build backend is missing the
  'build_editable' hook. Since it does not have a 'setup.py' nor
  a 'setup.cfg', it cannot be installed in editable mode. Consider using
  a build backend that supports PEP 660.

You might need to upgrade the dependencies in your local virtual
environment (re-run `python -m pip install -e .`).

We now need to depend on `pyproject.toml` too to cache dependencies in
the CI.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>