ci: Publish distribution files to PyPI
Add steps to publish distribution files to PyPI when a tag is created.

The Python version used to create the wheel is updated just for
consistency with other projects, it shouldn't really affect the
resulting wheel because this is a pure python package, so it doesn't
depend on the Python version, is just shipping text files.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>