Use the timestamp provided by the timer receiver
The resampling class now can take an optional timestamp to be used to
produce the new samples and the resampling actor now passes the
timestamp produced by the timer instead of calling `now()` manually.

This also adds a more explicit check to make sure the timer receiver is
not closed.

Fixes #54.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>