Minimize flakiness and simplify test
The resampling actor test was very complex, including starting a mock
gRPC server. This adds a lot of complexity to debug the test when it
fails and adds a lot of unnecessary complexity to the test and mocking
needs (we need to mock components in the component graph and interact
with the microgrid API when the resampling actor knows nothing about the
microgrid API).

To completely remove the flakiness, we need to use some solution to mock
time in asyncio (see #70).

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>