Don't ignore imports when running mypy
This allows for more strict type checking across dependencies
boundaries.

To be able to check types in a more strict way, we need to use packages
(-p) to avoid errors about having the same file in 2 packages
(src.frequenz.sdk and frequenz.sdk), so we need to translate
directory to package when running on `src`. Also we need to check
individual files separately.

A few type errors had to be fixed, most of them by adding type: ignore
because `grpc.aio` is missing stubs. Some other ignores were added to
tests for old data handling code that will probably be gone soon.

All ignores are in tests though, so it shouldn't be that bad.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>