Add support in FormulaEngine for handling `None` values
A new parameter `nones_are_zeros` is introduced in the `FormulaEngine`
and in the `FormulaBuilder`.

If this field is False,  then when there are None values in any of the
input streams,  the calculated values will be None.

When it is True, the None values will be treated as 0.0 when the
formula is evaluated.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>