Be more consistent when removing outdated samples
When removing outdated samples, samples falling just right at the
resapmling time are used in 2 different periods.

Even if this is not technically wrong, it is unexpected, as it would be
more natural that a moving window doesn't intersect with previous moving
windows (if using one period, or after moving the window
max_data_age_in_periods_s).

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>