Rewrite (`Group`)`Resampler` class
The `GroupResampler` and `Resampler` classes are not very reusable, or
at least require quite a bit of work from users to use them and they
need to take care of handling timers and calling methods at the right
time for them to be useful.

The new `Resampler` class has its own task(s) to handle all the
resampling, users only need to subscribe to new timeseries by passing a
timeseries `Source` and `Sink`. The `Sink` will be called each time a
new sample is produced.

A new `frequenz.sdk.timeseries.resampling` module is added because there
are more classes only used by the `Resampler` that might not be needed
by anyone just wanting to use `frequenz.sdk.timeseries`.

This commit also introduces the use of `async-solipsism` for testing, so
tests using timers can be mocked in a way that they don't really need to
wait for timers and time comparisons can be done precisely.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>