Rename resampling actor arguments
This is to make them easier to understand. Also require to use keyword
arguments to avoid mixing up `resampling_period_s` and
`max_data_age_in_periods`.

Also remove the example from the init documentation as it was outdated
and we already provide an example in `examples/` that at least gets
actually checked for syntax and linted.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>