Move resampler configuration to its own class
This makes user code a little bit more verbose but makes the code much
more maintainable, as we avoid having to copy and passing around a lot
of configuration variables that are only (or mostly) only relevant to
the internal resampling class.

It also removes a lot of documentation duplication that can get easily
out of sync and cause confusion.

And we'll add quite a few more configuration variables in subsequent
commits, which will just exacerbate the above mentioned issues without
a config class.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>