Use future annotations in resampling
Use built-in types and collections.abc types for type hinting as it can
be done on old Python versions when using:
    from __future__ import annotations

Also it seems like the type checker works better, because a cast is not
necessary anymore.

So imports from typing are needed still because this doesn't work for
type aliases, they still need stuff from typing.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>