Add a maximum resampling buffer size
This is mainly because in a near future the resampling buffer size will
dynamically change based on the input sampling rate, so we need to make
sure it doesn't get out of hand in the event of some unlikely high input
sampling rate combined with a very low frequency resampling.

We also add a warning size, for which we just emit a warning in the logs
this is to also try to minimize the presence of very big buffers without
affecting the running apps (unless again, it gets out of hand). When
this happens users should either try to use a different resampling
period to keep buffer smaller, or change the (default) warning buffer
length if they know it is still manageable.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>