Calculate the input sampling period
The input sampling period is calculated by counting the total received
samples, and dividing the total resampling time until the internal
buffer is full by the total received samples.

To store all the source properties (sampling period, sampling start,
total received samples) a new class is used (SourceProperties) and this
information is passed to the resampling function, so it can have more
information about the source to perform a proper resampling.

This calculation is performed once and then remains constant, but at
some point we could add a timer to re-calculate it.

This also improves slightly the documentation and validation of the
ResamplingConfig class.

The add_timeseries() methods now also takes a name as a way to identify
different sources in log messages (the actor uses the channel name as
timeseries name).

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>