Update the buffer length based on the input sampling period
When the input sampling period is know, we can adjust the internal
buffer to have a good size to be able to store all the requested
resampling and not waste space if the initial buffer was too big.

If we are upsampling, one sample could be enough for back-filling, but
we store max_data_age_in_periods using the input sampling period, so
the resampling functions can do more complex inter/extrapolation if they
need to.

If we are upsampling, we want a buffer that can hold
max_data_age_in_periods * resampling_period_s seconds of data, and we
have one sample every input_sampling_period_s, so we use a buffer length
of:
max_data_age_in_periods * resampling_period_s / input_sampling_period_s

The buffer size, like the input sampling period, is only calculated
once. The buffer data is copied to a new buffer when/if the buffer needs
resizing. The new size is validated against config.warn_buffer_len and
config.max_buffer_len and will emit a warning if it goes above any of
those and if it goes above config.max_buffer_len it will be truncated to
config.max_buffer_len as a safety measure.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>