Fix formula precedence bug in the `FormulaEngine`
Formulas with repeated operators like `#1 - #2 - #3` were getting
calculated incorrectly as `#1 - (#2 - #3)`.

The correct post-fix notation would be `[#1, #2, -, #3, -]`, but the
formula engine was generating this instead: `[#1, #2, #3, -, -]`.

This is fixed by flushing the operators from the build stack not only
if a previous operator has a higher precedence, but also if a previous
operator has the same precedence as the current operator.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>