Send 0 values from power formulas for non-existant component types
Earlier, the automatically generated formulas for `pv_power` and
`battery_power` were raising exception in cases where pv or battery
inverters were not found in the component graph, preventing us from
having a single high level formula that can be reused in all
locations.

This is solved by subscribing to a non-existing component id, so that
we can still send out 0 values at the same rate as the resampler.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>