Handle NaN/INF values in FormulaEngine
The MetricFetcher propagates NaN to handle None values
when None values are not treat from the stream as 0s.
So that any FormulaStep can compute the results without
checking for None on each value involved in the FormulaStep.
However the final result is written as None rather than NaN/INF
in the FormulaEngine.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>