Add MovingWindow
A data window that moves with the latest datapoints of a data stream.

When dealing with machine learning models, often the input features
are derived from timeseries data or are timeseries itself.

In order to obtain a timeseries as a data batch from streamed data
one can define a window that is essentially.

This approach creates a window that moves with the data flow by
utilizing a ringbuffer such that copying of buffer and therefore
memory allocations are minimized.

Signed-off-by: Matthias Wende <matthias.wende@frequenz.com>