Convert `FormulaChannel` into an ABC and specialize it for `Sample`s
This is achieved by converting the original implementations of the
following classes into Abstract Base Classes that are then derived
from the base class with the same name, for use with `Sample` and
`FormulaEngine` objects.

 - `FormulaChannel` -> `_BaseFormulaChannel` -> `FormulaChannel`
 - `FormulaReceiver` -> `_BaseFormulaReceiver` -> `FormulaReceiver`
 - `HigherOrderFormulaBuilder` -> `_BaseHOFormulaBuilder` -> `HigherOrderFormulaBuilder`

The will be specialized for 3-phase data also,  after a 3-phase formula
engine implementation is added.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>