Fix minimum requirements test
The test using minimum-requirements-ci.txt was broken.
The minimum-requirements-ci.txt file was not being used in
isolation because initially`all the dependencies were installed
from the pyproject.toml file. So dependencies that were present
in pyproject.toml but not in minimum-requirements-ci.txt won't
have triggered an error when testing the minimum requirements.

This patch generated a minimum dependencies list out of the dependencies
defined in pyproject.toml file and uses it instead of
minimum-requirements-ci.txt to avoid having differences between two
different sources of minimum requirements/dependencies.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>