Ensure all streams get synchronized in `FormulaEngine`
In cases where there are multiple streams lagging by the same number
of steps, only the one processed the last was getting synchronized to
the match the latest timestamp.

With this commit, `metrics_by_ts` stores a list of names for each
timestamp, so that each of them can get synchronized to the latest.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>