Move dev dependencies to pyproject.toml
With this we have only one central place where dependencies are
declared, so we can make sure we have a consistent view on the
dependencies.

As a side effect, the `formatting` session will install some extra
unneded dependencies (the package dependencies), but this allow us to
pin the tools versions and be sure everyone will use the same version.

A new alias is also added, so users wanting to setup a local development
enviroment can now just do a `pip install .[dev]` and that's it.

The CONTRIBUTING guide is updated to reflect this.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>