nox: Fix bug when invoking mypy
When invoking mypy, the src directory (frequenz.sdk package) isn't
being checked because 2 bugs:

* The "src" path is excluded
* A transformation of src into a package name was a ill done copy&paste
  from the channels repository, and said frequenz.channels instead of
  frequenz.sdk

This commit fixes this issue by reworking how the conversion between
paths and packages needed by mypy is done. Now the conversion is more
generic and explicit, providing a list of paths and a mapping between
paths and packages.

This is a bit more verbose to declare but it can be reused more safely
and gives some extra flexibility, which can be handy when we move this
to a general tooling repo.

We also use package names for plain python files (like noxfile.py), so
we don't need to invoke mypy twice.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>