Workaround some mypy issues
There seem to be 2 different mypy bugs at play while forwarding the call
to __setitem__().

First we need to check that the combination of input arguments are
correct to make the type checker happy (I guess it could be inferred
from the @overloads, but it's not currently working without this hack).

Then we need to ignore a no-untyped-call error, for some reason it
can't get the type for self._buffer.__setitem__().

We need to use `SupportsFloat` to handle cases where an `int` is used
instead of a float, that's used in the unit tests. Apparently `int` the
type checker accepts an `int` where a `float` is required but checking
`ininstance(value, float)` fails if `value` is an `int`.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>