Deal with missing type info in microgrid API calls
grpc.aio doesn't have type information, and there are all sorts of
tricks in it, so some funtions that look like they are not async really
are (because they return a awaitable). For example [`unary_unary()`]
returns a [`UnaryUnaryMultiCallable`] which returns a [`UnaryUnaryCall`]
which is `awaitable`.

Also there seem to be wrong type info regarding `SetBounds()`, saying
`request_iterator` is a required argument but it is not.
[`stream_unary()`] returns a [`StreamUnaryMultiCallable`] which takes an
optional `request_iterator`.

[`unary_unary()`]: https://github.com/grpc/grpc/blob/v1.51.3/src/python/grpcio/grpc/aio/_base_channel.py#L271-L276
[`UnaryUnaryMultiCallable`]: https://github.com/grpc/grpc/blob/v1.51.3/src/python/grpcio/grpc/aio/_base_channel.py#L28-L41
[`UnaryUnaryCall`]: https://github.com/grpc/grpc/blob/v1.51.3/src/python/grpcio/grpc/aio/_base_call.py#L138-L149
['stream_unary()']: https://github.com/grpc/grpc/blob/v1.51.3/src/python/grpcio/grpc/aio/_base_channel.py#L313-L318
[`StreamUnaryMultiCallable`]: https://github.com/grpc/grpc/blob/v1.51.3/src/python/grpcio/grpc/aio/_base_channel.py#L106-L118

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>