Exclude _formula_engine from type checking
This is to be able to merge the easy fixes until we have time to fix
this package, as it is quite complicated.

For this to work, we also need to install the package as editable for
`the ci_checks_max` nox session, otherwise the exclusion rule doesn't
work as it is not found in `src/`.

We also take the opportunity to make all installs in nox editable, it is
usually what we want while developing.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>