Add wrappers for the ComponentData in Unit tests
In the unit tests we have to mock component data with only few parameters.
Current ComponentData definition doesn't allow for it because
we have to give all class parameters.
We used to avoid that by creating protobuf message with only
necessary data and use it to create ComponentData.
But this approach is complex and not resilient for the changes in the
protobuf api.
This commit wraps ComponentData. Each parameter has default protobuf
value that will be used if no argument was given.

Signed-off-by: ela-kotulska-frequenz <elzbieta.kotulska@frequenz.com>