Fix flaky LogicalMeter test_soc
In MockMicrogrid it is hardcoded that iverter with id > 100
send first 5 messages and then stop sending.
Tests assumed we have control over when the components
starts sending data and in what order.
But that is not true in async.
Also inverter messages were not synchronized with rest of the data.

Signed-off-by: ela-kotulska-frequenz <elzbieta.kotulska@frequenz.com>