Make resampling window constant
The resampling window depends on when the resampling timer was fired.
This is wrong, because it makes the resampling window not constant, as
in general the timer will fire a bit after the resampling period has
passed, so it tends to be a bit bigger than resampling period.

This commit fixes this and uses a constant resampling window, always
emitting new samples with an exact multiple of the resampling period.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>