Don't include future samples in resampling windows
Currently only too-old samples are being filtered out from the relevant
samples by the resampler, but we should also filter out future samples,
as it might happen that the timer was delayed and was triggered a while
after a resampling period has passed, so we could have received samples
in the future for the current window's perspective.

This commit also filters these future samples and only consider relevant
samples that were emitted during the current resampling period.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>