Change PowerDistributor to ignore components with NaN metrics
Each float metric returned from api can be float("NaN").
If that happen PowerDistributor would fail because it is
impossible to make math calcuation on NaN value.
This PR is just a quick fix of this problem.

* if power bounds are NaN, then it tries to replace them
with corresponding power bounds from adjacent component.
If these components are also NaN, then it ignores battery.

* if other metrics metrics are NaN then it ignores battery.

More sophisticated solution should be done in the future.

Signed-off-by: ela-kotulska-frequenz <elzbieta.kotulska@frequenz.com>