Change BatteryPool to send metrics only when they change
We compared the metrics result but we included timestamp.
Because timestamps were different comparison was always True.
Change tells that timestamp should not be included in comparison
methods.
Remove Sample from the metric result types, because we don't
return any Sample metrics and it is not consistent with the rest
the result types, now.

Signed-off-by: ela-kotulska-frequenz <elzbieta.kotulska@frequenz.com>