Update MovingWindow size parameter type to timedelta
Modify the MovingWindow class to accept the size parameter
as a timedelta object instead of an integer. This change
allows the size parameter to represent the time span of
the moving window over which samples will be stored,
making it more intuitive for users and consistent among
other parameters in the MovingWindow class.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>