Add timezone to datetime types
There were inconsistencies in the way the SDK set
datetime variables in the MovingWindow/RingBuffer without
tz/tzinfo and in the Resampler with tz/tzinfo.
This patch sets the timezone UTC for all datetime variables
set in the MovingWindow and RingBuffer to make them consistent
to the datetime variables set in the Resampler.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>