Make power distributor's Results dataclasses
The Result class is an ABC but without any abtractmethods, a lot of
boilerplate can be removed by converting it to a dataclass and the same
for all subclasses.

This also uses mixins to reduce code duplication even further and making
sure result classes use consistent property names where it makes sense.

We still need to keep `Result` a class, but after the minimum supported
Python version is 3.10+ we can make `Result` a type union, so there is
not way to instantiate if by mistake, but still we can use
`isinstance(result, Success)` for example to what concrete type a result
it is, or even better use the `match` syntax.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>