Add an option to align resampling windows
The `ResamplerConfig` now has a new option `align_to` to align the
resampling windows to some arbitrary time. This means that the
timestamps of the produced samples will be a multiple of the resampling
period beginning at the `align_to` time.

By default `align_to` is the Unix epoch, and if it is `None` it will
have the current behaviour of aligning timestamps to the time when the
resampler is started.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>