Fix indexing in MovingWindow
When using a index slice where either the start or the stop key haven't
been defined these have been interpreted as None, which is the default
in the slice notation.

Instead using syntax where one is not set should be interpreted as an
interval starting with the first element or ending with the last
respectively.

Signed-off-by: Matthias Wende <matthias.wende@frequenz.com>