ci: Fix checks for merge queues
We now exclude testing the branches `gh-readonly-queue/**` as these are
the pushes GitHub does when "adding" PRs to a merge queue. We only want
to test the merge commit of a queue, to make sure all PRs included in
that merge play well together.

With this we can also remove the (wrong) hacky `if` conditions in
`test-generate-docs` and `publish-docs`.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>