Fetch PV power from PV meters, rather than PV inverters
This is because the microgrid API currently doesn't have drivers for
reading data from PV inverters.  We have plans to update our generated
formulas to have fallback components, when a certain selection of
components is unable to stream data, tracked here: #184.

Reading PV power from a meter is also going to be more accurate,
because it measures the actual power coming out of the PV inverter,
unaffected by the inverter's self consumption.

Signed-off-by: Sahas Subramanian <sahas.subramanian@frequenz.com>