Remove Sample comparison hack
The Sample object implemented comparisons by only comparing timestamps.
This was only done to cope with a limitation of `bisect()` before Python
3.10 (where no key argument could be passed).

Now that we are using Python 3.11 as the minimal version, we can remove
that hack and make Samples compare using both the datetime and the value
as one would expect.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>