Remove code providing multi-user support in PowerDistributor
Because the PowerDistributor doesn't have a way to resolve conflicts,
it can't be safely used by multiple users. So this feature will never
be used.

Once the PowerArbitrator is available, it will support multiple users
and will be the only component that will communicate with the
PowerDistributor.

This is being done now, in order to make the BatteryPool control
interface to be a bit more efficient and easy to implement.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>