Update PowerDistributingActor channel interface
Because multi-user support is going away, we can switch to a single
input receiver in the PowerDistributingActor for receiving requests.
So we don't need Bidirectional channels anymore.

We can't have a single output channel, because there would be multiple
request senders (possibly multiple EV Charger Pool instances, with
different component ID sets), and they would need to receive responses
just for the requests they made.

To rephrase, we are removing explicit multi-user support, and
replacing it with implicit multi-user support, which can also go away
once we have the PowerManager.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>