Take custom constructors in FormulaEngine for creating quantity objects
This commit also updates the FormulaBuilders to take custom
constructors as arguments.

FormulaEngines were previously taking the type of the quantity, with
which, it could use only the default constructors to create objects.

Because support for using default constructors is going away,
FormulaEngines and FormulaBuilders need to know the exact constructor
to use for creating output quantities.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>