Store a reference to the engine in receiver objects
This is a hack to ensure that the lifetime of the engine is tied to the
lifetime of the receiver.  This is necessary because the engine is a task that
runs forever, and in cases where higher order built for example with the below
idiom, the user would hold no references to the engine and it could get
garbage collected before the receiver.

    formula = (grid_power_engine + bat_power_engine).build().new_receiver()

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>