Allow MockMicrogrid battery chains without meters
In most component graph configs, batteries would be attached to
inverters and the inverters would be meters.  But there might be some
cases where there are no meters,  and we'll have to read data from the
inverters.

This commit enables such test configs, that can be used in tests that
use the MockMicrogrid.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>