Ensure component data is stale after receiving a timer event
Because of the nature of async, it might happen that data is slightly
delayed, so a timer event gets triggered, but component data arrives
before the timer event arrives,  and in such cases, we end up going
into a bad state.

This issue is resolved in this PR, by checking that the component data
is actually late, every time a timer triggers.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>