Always wait for at least one resampling period
When starting the resampler, if the time is not aligned to `align_to`,
it might happen that the first resampling window is too small, so it
will produce a bad first sample.

This commit ensures that the first window includes at least all the
samples of a full window, so the first resampling will be performed at
least one period after it was started (one period plus the time needed
for alignment).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>