Remove dunder iadd, isub and imul methods for Quantity
Having them mutating the objects in-place is not the best option as
number are assumed to be immutable in Python.  If we make them mutable,
then bad things will happen if they are used as default arguments in
functions, or if we want to provide a zero singleton for example.

We also mention explicitly now that Quantities are immutable.

This partially reverts 1043360af7b040674de553b7d663fa02404e5c19 (the
tests are kept).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>