Remove absolute tolerance in floating-point comparison
The comparison of the max excess `take_from[1]` and `deficit` against
zero was initially done using absolute tolerance. However, as both
values are then compared against each other and known to be away from
zero, the absolute tolerance becomes unnecessary. The relative tolerance,
which is defined by default as 1e-09, is sufficient for this comparison.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>