Improve and add more actor tests
In particular we test that the new restart rules work as expected, and
we also check the logs produced to have extra certainty that what we
expect to happen is really happening.

The restart tests are done using the `run()` function because we expect
to move the restart logic there, so the tests are "forward-compatible".

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>