Properly propagate `CancelledError` in `wait()`
When a `BackgroundService.wait()` is used, it should also propage
`CancelledError`s, as is it not always normal that a background service
is cancelled unless `stop()` is called, which uses `cancel()` to stop
the tasks.

When `stop()` is used, we properly filter out all `CancelledError`s too,
as in that case we triggered the cancellation, so it is normal.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>