[ci] Publish docs only if installation test succeeds
The test-installation step failing means that it is not possible to
install this python package on all-architectures. It is best not to update
published docs for a version that does not satisfy the requirement of being
installable on all intended target architectures.

Note that the test-installation step runs only if the build step is
successful.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>