Make exception documentation more generic
This exception is being raised indirectly, and the function raising it
was updated and now it raises the exception in different conditions.

Keeping the documentation of both functions in sync is costly and having
outdated documentation is confusing, so we just make the exception
documentation more general, hoping that it keeps being truth for longer.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>