Add a module for grid connection points
Grid connection points are nodes in the component graph that abstract
grids in a microgrid. Such nodes have their own metadata, which include
their rated fuse limits. Microgrids with zero or one grid connection points
are accepted. Note that a microgrid may have zero grid connection points,
if it is configured as an island.

This commit adds a module exposing a representation of such grid connection
points. The module is named `frequenz.sdk.microgrid.grid`.

The `grid` module exposes a singleton object that represents the grid
connection point. It gets initialized when `microgrid.initialize()` is
called. It can be obtained by calling `microgrid.grid_connection.get()`.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>