Implement a hash method for `Component`
Objects of the `Component` type are used in `ComponentGraph` in sets. The
implicit hash method works there for now, but adding new members to the
`Component` class may make it unhashable in the future, if the added
members themselves are unhashable. Implementing an explicit hash method
here removes this burden from from future member types.

Component IDs are supposed to be unique in a microgrid, therefore hashing a
`Component` object just based on its component ID is enough.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>