Replace the `section-index` plugin with `navigation.indexes`
The `section-index` plugin has a bug[^1] that causes the navigation bar
to display a wrong text for the **API Reference** link. There is a fix
for it, but `mkdocs-material` already provides the feature natively[^2],
so we can just use it instead, as it is much less likely to have issues.

[^1]: https://github.com/oprypin/mkdocs-section-index/issues/18
[^2]: https://squidfunk.github.io/mkdocs-material/setup/setting-up-navigation/#section-index-pages

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>