Apply changes suggested by the tool pyupgrade
The tool `pyupgrade` was used to modernize the code base
given that frequenz-SDK is already using python 3.11
as minimum version.
The tool was mainly used to replace obsolete types.

Just as a reference the tool was run with the following shell command:

```sh
find ./src ./benchmarks ./docs ./examples \
    ./tests -name '*.py' -exec pyupgrade --py311-plus {} \;
```

Also black and isort were both run to format the code base after
running pyupgrade.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>