Add count_valid method to ring buffer
The current len magic returns the count of samples in the ring buffer
that are determined valid. This definition is not unambiguously obvious
to the caller. The current logic is therefore moved to its own more
descriptive method which in the first step is called by the len magic.
This prepares removal of the len magic in subsequent changes.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>