Enable `nox-cross-arch` test for push events
If we don't do this, the `publish-docs` will not run on pushes, so the
in-development documentation will not be published.

While cross-arch tests are generally executed in the merge queue on the
identical commit, activating them for push events provides an added
layer of security. This is especially important in scenarios where
emergency pushes might bypass the merge queue.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>