Don't guess missing power bounds from adjacent components
While the API allows those values to be unset, in practice this won't
ever happen.
The backend will mark any components with missing power bounds as broken
and will not use them, so it makes no sense for us to guess those bounds
as the component won't be accepted by the backend anyway.

Signed-off-by: Mathias L. Baumann <mathias.baumann@frequenz.com>