Mock `_add_bounds_tracker` in the tests for the PowerManagingActor
And use a mock `_system_bounds` value to use with the test.

The tests for the PowerManagingActor are really just testing the
algorithm and don't use the MockMicrogrid, and `_add_bounds_tracker`
will not be able to track system bounds, causing the tests to fail.

Higher level tests using the MockMicrogrid are done as part of the
tests for the control interface of the BatteryPool.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>