Replace `handle_proposal` to `get_target_power`
`get_target_power` returns a target power only if there is a new
target power.  If the target power hasn't changed since the last call,
it returns `None`.

Its `proposal` parameter is optional.  This allows us to call this
function when the system bounds have changed, when no new proposals
have been received, which helps ensure that the target power is always
within the latest bounds.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>