Update obsolete comment about networkx-stubs
The networkx-stubs package was already working as the issue
mentioned in the comment was fixed. However, the package
networkx-stubs was removed a while ago from Frequenz-SDK as
it was no longer maintained and it has been archived.
At the moment of writing, there are only outdated and
incomplete alternatives to networkx-stubs.
Besides that, the Frequenz-SDK is not heavily using networkx,
so it is not worth it to maintain our own networkx-stubs package
or to contribute to a third-party package to fix errors.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>