Ignore unknown battery bucket in the absence of system bounds
If a proposal arrives for a previously unknown battery bucket, and
there are no system bounds available for the bucket, then ignore the
proposal.

This commit also improves the code comments for the case where a
previous target power already exists for a given battery bucket,
because in that case, we bring the target power down to zero.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>