Improve target_power calculation and handling of overlapping bounds
When actors with different priorities have not completely-contained,
but overlapping bounds, use the intersection to find the target_power
closest to the requirements of the lowest priority actor.

This commit also adds some test cases and updates the test to run
directly on a `Matryoshka()` instance, rather than through a
`PowerManagingActor` instance.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>