Resend request if PowerDistributor returns PartialFailure
- When there is an issue with any of the batteries, the power
distributor would mark the battery as temporarily broken and update
the battery status stream.

- Changes to the battery status stream would trigger big changes to
the system bounds calculated in the battery pool.

- Changes to the system bounds would not affect the target power
unless it was very big to begin with, so we need to resend the target
power to the power distributor, whenever a partial failure is received
from the power distributor.

- In the next run, the power distributor won't use the broken battery
and will return `Success`.

- If all batteries are broken and the power distributor returns an
`Error` instead, then the bounds will go down to zero, and it will
force a target power to get sent out already.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>